+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: fakoukou <fakoukou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/24 13:40:30 by fakoukou          #+#    #+#             */
-/*   Updated: 2024/11/02 15:23:59 by fakoukou         ###   ########.fr       */
+/*   Updated: 2024/11/04 13:01:26 by fakoukou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned char	*ptr;
 	char			x;
-	unsigned int	i;
+	size_t	i;
 
 	ptr = b;
 	i = 0;
